// frontend/src/features/editor-runtime/context/types/EditorRefsTypes.ts
import * as THREE from 'three';
import { TransformControls } from 'three-stdlib';
import type { MutableRefObject } from 'react';
import { ThreeSetup } from '../../hooks/useOpenDriveUtils/useThreeSetup/types/useThreeSetupTypes';
import { OpenDriveMapInstance } from '../../types/editorTypes';

export interface EditorRefs {
  sceneRef: MutableRefObject<THREE.Scene | undefined>;
  cameraRef: MutableRefObject<THREE.PerspectiveCamera | undefined>;
  carMeshesRef: MutableRefObject<THREE.Mesh[]>;
  carQuaternionsRef: MutableRefObject<Map<string, THREE.Quaternion>>;
  transformControlsRef: MutableRefObject<TransformControls | null>;
  rsuRenderMeshesRef: MutableRefObject<THREE.Mesh[]>;
  rsuObjectMeshesRef: MutableRefObject<THREE.Mesh[]>;
  routePointMeshesRef: MutableRefObject<THREE.Mesh[][]>;
  pedestrianObjsRef: MutableRefObject<THREE.Mesh[]>;
  rsuMeshesRef: MutableRefObject<THREE.Mesh[]>;
  buildingMeshesRef: MutableRefObject<THREE.Object3D[]>;
  roadMeshRef: MutableRefObject<THREE.Mesh | null>;
  currentCarRef: MutableRefObject<string>;
  currentColorRef: MutableRefObject<string>;
  loadPointsRef: MutableRefObject<() => void>;
  loadRSURef: MutableRefObject<() => void>;
  pedestrianMeshesRef: MutableRefObject<THREE.Mesh[]>;
  isDraggingRef: MutableRefObject<boolean>;
  mountRef: MutableRefObject<HTMLDivElement | null>;
  modeRef: MutableRefObject<{
    isAddCarModeActive: boolean;
    isAddPointModeActive: boolean;
    isAddPedestrianModeActive: boolean;
    isAddedPoints: boolean;
  }>;
  threeRef: MutableRefObject<ThreeSetup | null>;
  odrMapRef: MutableRefObject<OpenDriveMapInstance | null>;
}